Fix self._warn bug
In case the origin network name was not provided for the openstack
importer, it would call self._warn which would result in an error as there
is no _warn method defined. Now, instead of calling _warn we're making a
progress update so not only we avoid the error, but the user also gets to
know what's going on.