Add 'python-memcached' requirement.
When set to use `memcached` as a caching system, `keystoneauthmiddleware`
directly imports a `memcached`-related submodule of `oslo_cache`.
The problem is that neither `oslo_cache`, nor the `dogpile` package it
uses (which is configurable with a number of backends) actually directly
require `python-memcached` by default, and thus it will not be installed
at all leading to an ImportError.