Skip faulty offline disks during Windows OSMount.
In some situations the Windows worker VMs may have non-migrated disks
attached which they cannot bring online (ex: the config drive or
hypervisort-default floppy drive).
Coriolis should try to ignore disks which cannot be brought online, as
there is a good chance that it's not the OS disk and thus the OSMorphing
process should complete successfully.