Add 'storage_mappings' field to Migration/Replica API.
The 'storage_mappings' field will be passed to the provider plugins in
order to determine what storage backends to use for each of the
migrated/replicated VM disks.

The current implementation simply appends the 'storage_mappings' to the
'destination_environment' options, though an overhaul of the provider
interfaces to support (and allow for ease of addition of new options)
should be performed in the long run.