Fix deadlock, convert info field
  * There was a deadlock in worker.rpc.server caused by a large
enough task_info, due to the max buffer size of named pipes on linux.
We were calling join() on the process before consuming the Queue().
The process would never finish, because it would hang on queue.put()
  * Migrated the info field to LONGBLOB due to the new size of
task_info. We are now saving the disk state from previous runs where
the replicator is involved