Added the ability to add custom branches for provider repos.
The two dict variables added to the build config sample file should have
the same format: the key should be the provider name, and the value
should represent the "custom" branch or repo username.
If a provider is not found in the repo name dict, it will default to "cloudbase". If a
provider cannot be found in the branch dict, it will default to "master".

Example:
provider_custom_branch_names: {aws: some-aws-feature-branch, oci: quick-bug-fix}
provider_custom_repo_names: {aws: someuser, oci: torvalds}