Add HTTP backup writer
Leverage a HTTP based backup writer. The HTTP backup writer has the
added benefit of requiring less resources for large chunks, as it
does not need to rely on struct.pack to generate the payload it
sends over the wire. Standard HTTP calls are made to a remote API.

It also does not require a live, SSH channel to be active, and
recovery from failure is more likely to succeed. Removing the reliance
on paramiko also means that we are no longer constrained by the fact
that encryption of data over SSH done by paramiko is single threaded
and limited by the power of a single core available to the coriolis
worker.