Fix _expand_archive()
In cases where System.IO.Compression.ZipFile is available on the
windows minion, if the destination folder exists, it will cause
an error.

Where this component is not available, if the target folder exists
and there are conflicting files inside it with what's inside the
archive, it will error out.

In both cases, ensuring that the destination folder does not exist
will fix the issue.

If the destination folder is the root folder or is within the
%WINDOWS% folder, we do not remove it.