Fix OpenSSL usage for pass generation
Latest releases of LTS Linux distributions (for example Ubuntu 18.04)
come with OpenSSL 1.1.1 and above.

We use `openssl rand 18 -base64` to generate a random password.
While this works without a problem on OpenSSL 1.0.1 (shipped
with older Linux distributions, ex: Ubuntu 16.04), it does not
work with OpenSSL 1.1.1, giving the following args parsing error:
```
Extra arguments given.
rand: Use -help for summary.
```

This commit fixes the above error with OpenSSL 1.1.1. The change
has been tested on OpenSSL 1.0.1, and it does not break anything.