Use config file for the web proxy
Instead of passing the endpoints' URLs via environments, we use
the config file mounted from the host.

This will simplify the workflow when the config needs to be
changed. Instead of rebuilding the container with new env
variables, we just change the config file and restart the
current web proxy container.