Abstract the installation of cloudbase init
The plan is to install Cloudbase-Init on all providers, regardless
of whether or not Cloudbase-init can actually be used on the target
cloud. The idea is to have a single way to run user scripts on the
migrated instance.

For clouds where Cloudbase-Init does not have a provider, and we use
is as a glorified script runner, we can set the NoCloud metadata
provider.