Exclude task info from result
task info currently holds replica state for cases where the
coriolis replicator is used. We need to exclude this field when
returning a large result set, as the replica state can be quite
large, and the serialization of that field can take up considerable
resources. Moreover, the information in that field is only
relevant to the worker, not to the user.