Streamline task/minion pool event/progress update handling.
The old implementation of conductor task and minion pool events/progress
updates required "client-side" accouting of event/progress update
indexes and ordering, leading to excessive RPC/DB calls.
Additionally, the computation of percentages was also handled
"client-side" and simply added as a static message to the progress
update.

The new update adds an 'index' field to the DB tables for
events/progress updates which auto-increments for each task/pool, and
switches to using the 'current_step' (previously functioned as 'index')
and 'total_step' fields (previously unused) to represent the
current/total progress which can later be used to derive percentages.