Add memory profiling options for Coriolis components.
This patch adds support for setting up memory profiling for the root
process of each container running a Coriolis component. This is
achieved by using Fil memory profiler, which is installed using pip
in the 'coriolis-common' container and used to run the afferent
Coriolis service in the ENTRYPOINT of each container.