Prevent conductor from double-deallocating minion machines.
When cancelling an Execution which is in 'AWAITING_MINION_ALLOCATIONS'
status, the Conductor issues an explicit deallocation call to the Minion
Manager which trips up its TaskFlow allocation flow for said minions.
Considering it is not [currently] possible to cancel taskflow tasks,
this patch makes the Conductor skip the double-deallocation if the
Execution is in 'AWAITING_MINION_ALLOCATIONS' status.
As a result, the Minion Manager will automatically clear the allocations
once its Flow is finished and the minion reporting back to the Conductor
is refused.