Mark unit test folders as Python packages
Running `nosetests3 coriolis/tests` will skip any nested folders which
are not marked with `__init__.py` file.

Tests can still be run on folders which are not marked as a Python
packaged, but they need to be explicitly given when running the
`nosetests3` command. Ex.: `nosetests3 coriolis/tests/conductor/rpc`