Do not cache transport in worker
The worker process forks via multiprocessing when executing
provider code. This means that the transport needs to be initialized
after forking, in order for the provider to be able to communicate
with the conductor.

This means that there will be one transport creater per each python
thread that gets spawned and tries to communicate via RPC.

For the rest of the services, each process will have one transport,
regardless of how many threads it spawns.