Setup coriolis-logger as host service
This patch sets up coriolis-logger service as a systemd service on the host
instead of inside of a docker container. This change is made due to a metal-hub
issue, where setting up coriolis-metal-hub container logging driver as syslog,
and the address to coriolis-logger's socket file, the metal-hub container would
not start upon an appliance reboot.