Move `configure-step-ca` script to `utils`
This patch moves the step-ca configuration script to `utils` directory, so it
can be run on startup by rc-local. This script configures the default
certificate duration to 10y, and sets the DNS name of the CA server to the
main IP address, so the snapshot agents can be set up using the appliance's
main IP address.