Remove LVM system device filters for RedHat OSMount Tools
System Device filters is a feature enabled by default on RHEL9-based images,
which limits the LVM detection on a particular set of system devices, provided
in the `system.devices` file. This is bad for the OSMount process, since it
will fail to detect LVM volumes on the migrated partitions. Removing the
`system.devices` file from the minion machine disables this feature and fixes
the issue.