Fix reservation management on minion pool errors
Fixes an edge-case where if a reservation cannot be found for an existing
replica, and the minion pool set to it is also in an invalid state, a new
reservation would be created but never saved to DB because of the minion pool
error. This patch makes sure that the minion pool validation is done before
the reservation checks when creating replica executions.