Fix transfer schedules not being able to delete after deployment execution
When a schedule with `auto_deploy` enabled triggers the transfer and pending
deployment using a user trust that's registered in the schedule information.

After a deployment finishes, the trust_id is always deleted, making it
impossible for the user to delete an auto_deploy-enabled transfer schedule.

This patch makes sure that a transfer schedule can be deleted, even though
its trust_id fails to delete.