Fix deployer manager failing deployments when awaiting for minions
If the deployer transfer starts with a minion pool, the deployer manager
would've considered the `AWAITING_MINION_ALLOCATIONS` status as an invalid
one, and thus error out the automatically launched deployment. This patch
adds this minion status as an active and valid one, and properly waits for
the deployer to finish with minion refreshment and the transfer execution as
well.